requirements.txt
django >3.2,<3.3
python-decouple
pillow
djangorestframework
markdown
django-filter
drf-yasg[validation]
# rest_framework.authtoken
djangorestframework-simplejwt
django-mail-templated

black
